Initial Commit
This is the first commit for the modernized skyrise code. This code only contains driver control code. This is the first multi-file program that I have written, so I expect this to be a bit of a learning process. The code has NOT yet been debugged.

Main.c: This is the main file. It contains the usercontrol and autonomous tasks, as well as the pre_auto function (though both the auto task and pre_auto function are empty). The user control section uses a number of tasks defined in the various header files.

Tele.c : This contains all of the headerfiles for the driver-control period.

Auto.c: Not yet used. this will store include statements for all autnonmous headerfiles.

LCD.h: This contains the LCD task. Copied from ITZ skills robot

Tele_Drive: This contains a joystick velocity control task and drive task. Both have been copied and slightly modified (to have lower memory load) from ITZ code.

Tele_Intake.h: I copied the old driver control code and put it in a task.

Tele_Skyrise.h: I copied the old skyrise and tail controls and put them in a new task. This has been cleaned up so that the skyrise, tail numbers should always be either 0 or 1. I also got rid of the Skyrise_Action_Number and Tail_Action_Number funtions in favor of a single task.

Tele_Lift.h: This is simply the lift task that was in my old skyrise code... has not been modified.