Fixed bugs, rewrote LCD
I fixed some bugs (mostly typos) in the initial commit. This code now runs on the robot!

I rewrote the majority of the LCD task, I wasn't a fan of the old way I created the display strings. I rewrote it in a way that made more sense to me.

I also had to move the left encoder onto port 5,6 (rather than 3,4). One of ports 3,4 is dead. This forced me to redo the ports for most of the digital sensors.