Fixed Auto_Drive, Drive_Assist errors. Robot can autonomously drive now
I fixed a number of syntax errors with the last version of the code. Most of these were minor typos not worth mentioning. However, there were a few serious syntax errors. For one, you can't initlaize a strcut? at least not the way I was doing it. I moved the assignment of the Turn_PD control strcuture into the Drive_Assit function (it gets set up before the function runs for the first time).

NOTE: THE TURN_PD STRCUT IN THE AUTO_DRIVE.H FILE MUST BE REMOVED WHEN I WRITE AN AUTO TURN TASK.

There was also a logical error in the Drive_Assist function. The correction to the left and right side of the drive was the negative of what it should have been. This stems from how I defined turning error in the ITZ code (really stupid). In general Error should be defined as follows: Target - Position. I plan on rewriting all loops using this rule. Once these corrections were made, the robot was able to drive (and keep straight!).

Drive, Drive_Assist control loops need considerable tuning!!!!!