Initial commit. Added debug printing
This branch is designed to measure how the robot oscillates with a purely P driving function.

To do this, I turned off the D part of the PD control. Further, I added a debug line that prints out the position and velocity of the robot at each iteration (the time is also printed). This data will be used to find out how the system naturally decays, and the period of oscillation. This will be used to determine the robot's 'natural mass' and 'natural damping'. This data will allow for better PD control with critical damping.

This code has been written but not debuged. This code NEEDS to be debuged. Do not use this version.