Added Power_Array generator, got oscillation data
I downloaded/adapted the Power_Array generator that I made last summer. This code needed to be modified to work with the skyrise drive. I also choose to run this test on a floor. I had to reduce the runtime to .1 seconds (down from .5) so that it could drive the full distance without driving into a wall. The data took a LOT of tries to get, but the new power array appears to work better. The final array is stored in the Skyrise_Power_Array text file.

Once I hada better power array, I got oscillation data with a 50 and 25 ms update time. I also reduced the kp to get better oscillations. Overall, the drive appears to display decaying oscillations for a few cycles, then get into a strange stalbe oscillation. I believe that first few oscillations is all that is useful.