Measured deg per tick (using LCD)
I measured the number of degrees per tick using the LCD readout. The theoretical value is 3.25/15 = ~ .217. The measured value was around .209... pretty similar. The old value was way off (I forget how it was calculuated, but it was wrong). I should now be able to write better tick based auto.