Redid PD_Control struct
I removed the Target_Sign member and added an Error and 'k' member. The error member is a 5 element array (the error array that use to be defined in each task). The k member is simply the error update index. The value of k tells us which element of the Error array is the newest.

I removed the Target_Sign member becaue Drive_Assist was the only task that was using it. I want members of the strcut to be variables that every task uses. As such, it didn't make sense to keep the Target_Sign inside of the strcut (just wastes memory).

Every task that used PD_Control used an Error array. I included the Error array and k members so that external tasks can get error information. One issue, however, is that the error array is updated in a cyclic manner. In order to know which error is the latest, we need to know which index was last updated. This is exactrly what the k member does. WIth this information, we can calculate the position (Target - error[k]), velocity (using 5 point forumla) and Error.

I choose to keep the local k index in most tasks simply because typing out Turn_PD.k or Drive_PD.k would make the code less readible. Thus, I kept the local k variable in these tasks as an alias for the structure k member.

With this information in the struct, the LCD function should be able to display veocity,  position, and error information in real time.