Updated power array generator code + got turning data
I massivly redid the power array generator code (PowerVsRPMTesT). Rather than having the robot drive forward, I had the robot drive in circles. this way I can have the robot run at each power setting for a full .5 seconds without needing a large area to test in (I can also test on field tiles!).

In addition to this change, I created a new definition called UPDATE_TIME (which does as the name would imply). Thus, the time for each power setting can be controlled with this definition. I also added in my classic Byte and uByte typedefs and used them where appropiate. Finally, I realized that I should probably print out the normalized RPM array in addition to the power array, just so that I can verify that the  program is working as intended. This meaned running the print code again. Rather than using the same code twice, I wrote a Print_Array function to do this. The printing code was moved to this function. The program calls this function twice, once to print the Normalized RPM array and once to print the Power Array.

Finally, I improved the printout. This was done by using a specified integer length(%3i, pads with up to 3 whitespaces), and moved the {, } to their own line + added a ; (and some new lines to make things work nicer). This makes the printout ahdere to the aray syntax I use in the actual skyrise/competition code.

With these changes, I tried to get a new power array. I ran into a small issue however. I noticed that in all my power arrays (like everyone I have ever made up until now) the first several elements are always the same value. I know why this is, static friction! The satic coefficient of friction tends to be higher than the kinetic coefficient of friction. As a result, it takes a lot of force to get the robot moving/overcome the static friction. However, once the robot begins moving, the robot only has to deal with kinetic friction. The robot therefore rapidly accelerates. This means that the RPM array has a bunch of zeros, then rapidly jumps to 20 or so in a few indicies (where the robot begins moving). As a result, the robot thinks that the same index corresponds to all rpm's between 0 and 20, thus the repated elements in the power array. To fix this, I literally pressed on the robot with my finger as soon as it started. I did this until the robot began moving. I tried to press lightly/just hard enough to make up the difference between static and kinetic friction. The result was a much better looking power array! This is stored in the Turn_Power_Array.txt file (along with a few of my mess up's... note, the first and every other array is an rpm array, the others are Power Arrays.. since both are printed each time I run the program)

With this improved power array, I tested the rotation test. Luckily, I was able to get a few oscillations/some good data. This data is stored in the 50ms_Turn_Data.txt file. As the name implies, I used a 50ms refresh time and a kp of 5.